+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notables change for @leomotors/music-bot will be noted here
 
+## [0.4.0] - 2022-05-06
+
+- The bot now tell user if it is attempting to joining empty or not joinable functions
+
+- Add /rejoin to force bot to join the voice channel
+
 ## [0.3.0] - 2022-05-03
 
 - Fixed an issue where all server share the same music state (Reference and Value Stuff, again)